Revise file casing to kebab case
