+++ components/hero-image.tsx
