+++ components/image.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import NextImage, { ImageProps } from 'next/image';
 import Link from 'next/link';
-import Wrapper from './Wrapper';
+import Wrapper from './wrapper';
 
 interface ImageComponentProps extends ImageProps {
   showCaption?: boolean;
