+++ components/seo.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Head from 'next/head';
 import { useRouter } from 'next/router';
-import { siteMetadata } from '@/data/siteMetadata';
+import { siteMetadata } from '@/data/site-metadata';
 import { CoreContent } from '@/lib/utils/contentlayer';
 import type { Blog, Authors } from 'contentlayer/generated';
 interface CommonSEOProps {
