+++ pages/_app.tsx
@@ -6,8 +6,8 @@ import { ThemeProvider } from 'next-themes';
 import type { AppProps } from 'next/app';
 import Head from 'next/head';
 
-import { siteMetadata } from '@/data/siteMetadata';
-import LayoutWrapper from '@/components/LayoutWrapper';
+import { siteMetadata } from '@/data/site-metadata';
+import LayoutWrapper from '@/components/layout-wrapper';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
