+++ pages/_document.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Document, { Html, Head, Main, NextScript } from 'next/document';
-import { siteMetadata } from '@/data/siteMetadata';
+import { siteMetadata } from '@/data/site-metadata';
 
 class MyDocument extends Document {
   render() {
