+++ pages/tags/[tag].tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { TagSEO } from '@/components/SEO';
-import { siteMetadata } from '@/data/siteMetadata';
-import ListLayout from '@/layouts/ListLayout';
-import { kebabCase } from '@/lib/utils/kebabCase';
+import { TagSEO } from '@/components/seo';
+import { siteMetadata } from '@/data/site-metadata';
+import ListLayout from '@/layouts/list-layout';
+import { kebabCase } from '@/lib/utils/kebab-case';
 import { getAllTags, allCoreContent } from '@/lib/utils/contentlayer';
 import { InferGetStaticPropsType } from 'next';
 import { allBlogs } from 'contentlayer/generated';
