+++ scripts/rss.ts
@@ -1,5 +1,5 @@
 import { generateRSS } from '../lib/utils/generate-rss';
-import { siteMetadata } from '../data/siteMetadata';
+import { siteMetadata } from '../data/site-metadata';
 import { allBlogs, allAuthors } from '../.contentlayer/generated/index.mjs';
 
 const rss = () => {
