+++ scripts/search.ts
@@ -1,7 +1,7 @@
 import { writeFileSync } from 'fs';
 import { allCoreContent } from '@/lib/utils/contentlayer.js';
 import { allBlogs } from '../.contentlayer/generated/index.mjs';
-import { siteMetadata } from '../data/siteMetadata.js';
+import { siteMetadata } from '../data/site-metadata.js';
 
 const search = () => {
   if (siteMetadata?.search?.kbarConfig?.searchDocumentsPath) {
